+++ src/classes/errors/index.ts
@@ -1,0 +1,4 @@
+export * from './delayed-error';
+export * from './unrecoverable-error';
+export * from './rate-limit-error';
+export * from './waiting-children-error';